cambios menores

+++ V2.0/clienteV2.0/conex/comandos/ss/ss.go
@@ -58,7 +58,7 @@ func Obtener_img(conn net.Conn) ([]byte, error) {
 	}
 	total := binary.BigEndian.Uint64(tamaño_img) //tamaño total de la imagen
 
-	fmt.Printf("cantidad total de la imagen: %d MB aprox\n", total/1_000_000)
+	fmt.Printf("cantidad total de la imagen: %f MB aprox\n", float32(total)/1_000_000)
 
 	img := make([]byte, total) // crear buffer de la imagen
